Improved help
Moved help related functions to help.go.
Added heading modifiers.
Fixed the heading item.
