+++ go.mod
@@ -11,6 +11,7 @@ require (
 	github.com/fsnotify/fsnotify v1.5.1
 	github.com/gdamore/tcell/v2 v2.4.0
 	github.com/golang/glog v1.0.0 // indirect
+	github.com/jwalton/gchalk v1.2.1
 	github.com/klauspost/compress v1.13.6
 	github.com/mattn/go-runewidth v0.0.13
 	github.com/pierrec/lz4 v2.6.1+incompatible
@@ -19,7 +20,8 @@ require (
 	github.com/spf13/viper v1.9.0
 	github.com/ulikunitz/xz v0.5.10
 	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c
-	golang.org/x/sys v0.0.0-20211103235746-7861aae1554b // indirect
+	golang.org/x/sys v0.0.0-20211113001501-0c823b97ae02 // indirect
 	golang.org/x/term v0.0.0-20210927222741-03fcf44c2211
 	golang.org/x/text v0.3.7 // indirect
+	gopkg.in/ini.v1 v1.64.0 // indirect
 )
